requirements.txt
numpy>=1.19.2
Pillow>=8.0.0
imagehash>=4.2.1
scipy>=1.5.2
tqdm>=4.65.0
annoy>=1.17.1
fire>=0.5.0